Fix: stop using CID as ref for STORE messages
Problem: using a CID as reference is now taken into account by
the message processing dependency resolution system and interpreted
as the hash of a STORE message to update. As we use the CID of the file,
a corresponding aleph.im message never exists and the message is
discarded.

Solution: set the ref field to "request-<cid>".